(fix) Fix event details view from feed
