+++ .eslintrc
@@ -16,6 +16,7 @@
     "react/forbid-prop-types": 0,
     "react/jsx-filename-extension": 0,
     "react/jsx-no-bind": 0,
-    "no-underscore-dangle": 0
+    "no-underscore-dangle": 0,
+    "class-methods-use-this": 0,
   }
 }
